refactor: move lib to utils
