+++ src/domain/types.rs
@@ -3,7 +3,7 @@ use nanoid::nanoid;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::str::FromStr;
 
-use crate::lib::error::{ApplicationError, ApplicationResult};
+use crate::utils::error::{ApplicationError, ApplicationResult};
 
 // TODO: Value Object用のderive macroを作る
 // ↓みたいな一要素のタプル構造体たちにfrom, valueをデフォルトで実装したい
