+++ src/infrastructure/dto/coordinate.rs
@@ -4,7 +4,7 @@ use crate::domain::coordinate::Coordinate;
 use crate::domain::types::RouteId;
 use crate::infrastructure::dto::route::RouteDto;
 use crate::infrastructure::schema::coordinates;
-use crate::lib::error::ApplicationResult;
+use crate::utils::error::ApplicationResult;
 
 /// 座標のdto構造体
 #[derive(Identifiable, Queryable, Insertable, Associations, Debug)]
